Removed AnnotationDAO interface
